WIP: we do not seem to have memory leaks anymore, however, we are now 5x slower than vanilla spark (again) -- time for timings! -- making an 'compiled query'-sorter -- does not make much of a difference: perhaps use timsort?

+++ arrowspark/src/main/scala/org/apache/spark/sql/column/ArrowColumnarBatchRow.scala
@@ -26,7 +26,7 @@ import java.io._
 class ArrowColumnarBatchRow(@transient val allocator: BufferAllocator, @transient protected[column] val columns: Array[ArrowColumnVector], val numRows: Int) extends InternalRow with Closeable {
   override def numFields: Int = columns.length
 
-  def length: Long = numRows
+  def length: Int = numRows
 
   override def isNullAt(ordinal: Int): Boolean = ordinal < 0 || ordinal >= numFields
